simplified StringBuf_push_string
Removed the __builtin_clz nonsense because it's hard to get right and
premature optimisation is bad, also I'd rather this project be as
portable as possible. I could've used Window's _BitScanReverse to help
with portability but I tried that out and couldn't get it to work
because Microsoft is dumb and weird.